+++ src/print/unix/mod.rs
@@ -1,6 +1,7 @@
 use std::{borrow, ffi};
 
 pub mod cups;
+pub mod dest;
 pub mod job;
 pub mod native;
 pub mod options;
